requirements.txt
streamlit>=1.55
langchain>=1.2
langchain-community>=0.4
langchain-google-genai>=4.2
langchain-text-splitters>=1.1
faiss-cpu>=1.13
PyPDF2>=3.0
python-dotenv>=1.0
tiktoken>=0.12